# api.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SMARTBEARD192M/FF-ICE_Demo/1.0.1
info:
  description: API for FF-ICE Demo App
  version: "1.0.1"
  title: FF-ICE Demo API
  contact:
    email: a.arango@tu-braunschweig.de
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  /routes:
    get:
      description: "All flight routes"
      parameters: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightRoute"
      summary: Get all routes
      tags:
        - default
    post:
      description: Save a new route
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightRoute"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  insertedId:
                    type: string
                    description: Id of the inserted record
components:
  schemas:
    Location:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          minimum: -90
          maximum: 90
          description: "Latitude"
        longitude:
          type: number
          minimum: -180
          maximum: 180
          description: "Longitude"
        altitude:
          type: number
          minimum: -5000
          maximum: 200000
          description: "Altitude in ft"
    Waypoint:
      type: object
      required:
        - waypointId
      properties:
        waypointId:
          type: string
          description: "Unique waypoint id"
          maxLength: 5
          minLength: 3
        location:
          $ref: "#/components/schemas/Location"
    FlightRoute:
      type: object
      required:
        - routeId
      properties:
        routeId:
          type: string
          description: Route id
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/Waypoint"
...